pyproject.toml
[tool.poetry]
name = "warehouse-service-test-task"
version = "0.1.0"
package-mode = false
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.116.1"}
pytest = "^8.4.1"
httpx = "^0.28.1"
pytest-asyncio = "^1.1.0"
pytest-fastapi = "^0.1.0"
uvicorn = "^0.35.0"
aiokafka = "^0.12.0"
sqlalchemy = "^2.0.42"
pydantic = "^2.11.7"
alembic = "^1.16.4"
asyncpg = "^0.30.0"
ruff = "^0.12.8"
isort = "^6.0.1"
black = "^25.1.0"
fastapi-cache2 = "^0.2.2"
redis = "^6.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
